requirements.txt
pandas
numpy
faiss-cpu
sentence-transformers
crewai
langchain-community
ollama